requirements.txt
streamlit
chromadb
pandas
google-generativeai
python-dotenv
